gwy: file post srlzr now using common path sanitization function
